fix 1 for v9.2-2

+++ Lib/distutils/openvmsccompiler.py
@@ -158,7 +158,7 @@ class OpenVMSCCompiler(CCompiler):
         if debug:
             cc_args += ["/DEBUG/NOOPTIMIZE/LIST/SHOW=EXPANSION"]
         else:
-            cc_args += ["/NODEBUG/OPTIMIZE"]
+            cc_args += ["/NODEBUG/NOOPTIMIZE"]
         if VMS64:
             cc_args += ["/POINTER_SIZE=64"]
         
@@ -325,7 +325,8 @@ class OpenVMSCCompiler(CCompiler):
             except DistutilsExecError as msg:
                 raise LinkError(msg)
             finally:
-                os.unlink(opt_file.name)
+                if not os.getenv("PYTH$LINK_NO_DEL_OPT", None):
+                    os.unlink(opt_file.name)
         else:
             log.debug("skipping %s (up-to-date)", output_filename)
